chore(dev playground): clean up dev playground for better experience
